+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
-## Unreleased
+## 1.4.1
 
 ### Fixed
 
 - `Stream::create` with a string needs to rewind the created memory stream.
+- `Psr17Factory::createStreamFromFile`, `UploadedFile::moveTo`, and
+  `UploadedFile::getStream` no longer throw `ValueError` in PHP 8.
 
 ## 1.4.0
 
@@ -118,4 +120,3 @@ The `final` keyword was replaced by `@final` annotation.
 ## 0.2.3
 
 No changelog before this release
-